requirements.txt
asgiref==3.7.2
boto3==1.34.51
botocore==1.34.51
cachetools==5.3.3
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
google-auth==2.28.1
idna==3.6
jmespath==1.0.1
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
uvicorn[standard]
